// src/components/calculator/types.ts
export interface PropertyDetails {
  sellerName: string;
  propertyAddress: string;
  purchasePrice: number;
  taxesApprox: number;
  docStampsDeed: number;
  ownersTitleInsurance: number;
  commissionRate: number;
  commission: number;
  complianceAudit: number;
  serviceTech: number;
  termiteInspection: number;
  fhaVaFees: number;
  survey: number;
  hoa: number;
  homeWarranty: number;
  buyersClosingCost: number;
  repairs: number;
  searchExamClosingFee: number;
  sellerPayingTitle: boolean;
  firstMortgage: number;
  secondMortgage: number;
}